+++ SAM.swiftpm/Model/Bitmasks.swift
@@ -1,0 +1,15 @@
+//
+//  File.swift
+//  
+//
+//  Created by 077tech on 2023/03/28.
+//
+
+import Foundation
+
+enum bitmasks: UInt32 {
+    case frame = 0b1
+    case paddel = 0b10
+    case stone = 0b100
+    case ball = 0b1000
+}
