+++ SAM.swiftpm/Resources/Assets.xcassets/Background/background.imageset/Contents.json
