+++ SAM.swiftpm/Resources/Assets.xcassets/Blocks&Paddel/paddel.imageset/Contents.json
