+++ SAM.swiftpm/View/GameScene2.swift
@@ -15,13 +15,6 @@ class GameScene2: SKScene, SKPhysicsContactDelegate {
     let paddel = SKSpriteNode(imageNamed: "paddel")
     let ball = SKSpriteNode(imageNamed: "ball")
     
-    enum bitmasks: UInt32 {
-        case frame = 0b1
-        case paddel = 0b10
-        case stone = 0b100
-        case ball = 0b1000
-    }
-    
     override func didMove(to view: SKView) {
         scene?.size = view.bounds.size
         scene?.scaleMode = .aspectFill
