+++ SAM.swiftpm/View/GameScene3.swift
@@ -14,13 +14,6 @@ class GameScene3: SKScene, SKPhysicsContactDelegate {
     let paddel = SKSpriteNode(imageNamed: "paddel")
     let ball = SKSpriteNode(imageNamed: "ball")
     
-    enum bitmasks: UInt32 {
-        case frame = 0b1
-        case paddel = 0b10
-        case stone = 0b100
-        case ball = 0b1000
-    }
-    
     override func didMove(to view: SKView) {
         scene?.size = view.bounds.size
         scene?.scaleMode = .aspectFill
