+++ SAM.swiftpm/View/OnBoardingPage.swift
@@ -1,0 +1,15 @@
+//
+//  SwiftUIView.swift
+//  
+//
+//  Created by 077tech on 2023/03/28.
+//
+
+import SwiftUI
+
+struct OnBoardingPage: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
