+++ SAM.swiftpm/View/StoryScene1.swift
@@ -1,0 +1,16 @@
+//
+//  SwiftUIView.swift
+//  
+//
+//  Created by 077tech on 2023/03/28.
+//
+
+import SwiftUI
+
+struct StoryScene1: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+
