+++ SAM.swiftpm/View/StoryScene2.swift
@@ -1,0 +1,14 @@
+//
+//  SwiftUIView.swift
+//  SAM
+//
+//  Created by 077tech on 2023/03/28.
+//
+
+import SwiftUI
+
+struct StoryScene2: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
