+++ SAM.swiftpm/View/StoryScene4.swift
@@ -1,0 +1,15 @@
+//
+//  SwiftUIView.swift
+//  
+//
+//  Created by 077tech on 2023/03/28.
+//
+
+import SwiftUI
+
+struct StoryScene4: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
